requirements.txt
requests
pyyaml
Pillow
pytz
mysql-connector-python
psycopg2-binary